Lib

Documentation

data Playground

Constructors

Playground

Fields

normalMatrix :: Matrix Int

Matrix used for filling and playing.

controlMatrix :: Matrix Int

Matrix used to validate the Solution.

field :: Field

Field that was most recently used.

startField :: Field

The Starting Field of the Game.

size :: Int

Size of the Playground.

solved :: Bool

Indicates if the last checked solution is correct.

data Field

Constructors

Field

Fields

val :: Int

Value of the field.

pos :: (Int, Int)

Position of the field.

getMatrix :: Playground -> Matrix Int

Getter for the matrix of the playground.

getControlMatrix :: Playground -> Matrix Int

Getter for the control matrix of the playground.

getField :: Playground -> Field

Getter for the field of the playground.

getStartField :: Playground -> Field

Getter for the start field of the playground.

getSize :: Playground -> Int

Getter for the size of the playground.

getStatus :: Playground -> Bool

Getter for the solved status of the playground.

getValue :: Field -> Int

Getter for the value of the field.

getPosition :: Field -> (Int, Int)

Getter for the position of the field.

generateEmptyMatrix

Arguments

:: Int
Size of the matrix.

-> Matrix Int
The zero matrix.

Creates a zero Matrix.

setStartPoint

Arguments

:: Matrix Int
Matrix to put the start point in.

-> (Int, Int)
Position for the start point.

-> Int
Size of the Playground.

-> Playground
The playground with a set start point.

Sets the start point.

calcNeighbours

Arguments

:: Matrix Int
Matrix to search in.

-> (Int, Int)
Position of the field.

-> Int
Size of the playground.

-> [Field]
Returns a field list.

Calculates the neighbours of the given position with a 0 value.

upperNeighbour

Arguments

:: Matrix Int
Matrix to search in.

-> (Int, Int)
Position of the field

-> Int
Size of the playground for range check.

-> Field
returns a field.

Calculates if there is a valid upper neighbour.

lowerNeighbour

Arguments

:: Matrix Int
Matrix to search in.

-> (Int, Int)
Position of the field

-> Int
Size of the playground for range check.

-> Field
returns a field.

Calculates if there is a valid lower neighbour.

leftNeighbour

Arguments

:: Matrix Int
Matrix to search in.

-> (Int, Int)
Position of the field

-> Int
Size of the playground for range check.

-> Field
returns a field.

Calculates if there is a valid left neighbour.

rightNeighbour

Arguments

:: Matrix Int
Matrix to search in.

-> (Int, Int)
Position of the field

-> Int
Size of the playground for range check.

-> Field
returns a field.

Calculates if there is a valid right neighbour.

upperLeftNeighbour

Arguments

:: Matrix Int
Matrix to search in.

-> (Int, Int)
Position of the field

-> Int
Size of the playground for range check.

-> Field
returns a field.

Calculates if there is a valid upper left neighbour.

upperRightNeighbour

Arguments

:: Matrix Int
Matrix to search in.

-> (Int, Int)
Position of the field

-> Int
Size of the playground for range check.

-> Field
returns a field.

Calculates if there is a valid upper right neighbour.

lowerLeftNeighbour

Arguments

:: Matrix Int
Matrix to search in.

-> (Int, Int)
Position of the field

-> Int
Size of the playground for range check.

-> Field
returns a field.

Calculates if there is a valid lower left neighbour.

lowerRightNeighbour

Arguments

:: Matrix Int
Matrix to search in.

-> (Int, Int)
Position of the field

-> Int
Size of the playground for range check.

-> Field
returns a field.

Calculates if there is a valid lower right neighbour.

setPoint

Arguments

:: Playground
Playground to enter the point in.

-> Int
Value to enter.

-> IO Playground
Playground to return.

Sets the next point into the playground. Uses a shuffle function to randomise the position of the field.

decider

Arguments

:: Playground
Playground.

-> Matrix Int
Matrix.

-> Field
Field.

-> [Field]
Neighbours list.

-> Playground
Playground to return.

Used for checking if there a is neighbour nearby.

shuffleList

Arguments

:: [a]
Input list.

-> IO [a]
Return shuffled list.

Function used to shuffle a given list.

returnFieldFromNeighbours

Arguments

:: [a]
List.

-> IO a
Returns the first value of the shuffled list.

Shuffles a list and returns the first value.

buildPlayground

Arguments

:: Int
Size of the playground.

-> IO Playground
Returns filled playground.

Used to initiate the playground and fill it.

setPoints

Arguments

:: Playground
Playground to enter the points.

-> [Int]
List of points to enter.

-> IO Playground
Returns the playground with set points.

Used for entering all points into the playground.

generatePlayground

Arguments

:: Int
Size of the playground.

-> IO Playground
Returns the built playground.

Builds a playground and checks if it fits the wanted fill rate.

zeroRate

Arguments

:: Matrix Int
Matrix.

-> Float
Returns the percentage as float.

Calculates the percentages of empty to filled fields of a matrix.

generatePlaygroundToPlayOn

Arguments

:: Playground
Playground with all points set.

-> IO Playground
Playground with removed points for playing.

Takes a playground with all set points and removes some of them, so that the playground is ready for playing.

gameHandler

Arguments

:: Playground
Playground to play on.

-> IO ()

Used for handling the input from the user and checks if game finished(gameloop).

handler

Arguments

:: Playground
Playground.

-> String
Entered input from user.

-> Playground
Returns the playground.

Handles the input from the user and calls different functions depending on the input.

enterValue

Arguments

:: Playground
Playground.

-> Int
Value to enter.

-> (Int, Int)
Position to enter the value in.

-> Playground
Return the playground with entered value.

Function to enter a value into the playgroung. Checks if Position is valid, so no given values are overwritten.

checkSolution

Arguments

:: Playground
Playground.

-> Field
Start field of the playground.

-> Playground
Returns playground with either True or False as Status solved.

Function used for validating the solution from the user.

calcSolutionNeighbours

Arguments

:: Matrix Int
Matrix to search in.

-> Field
Given field to search around.

-> Int
Size of the Playground.

-> [Field]
Returns the neighbour as list.

Calculates if there is the correct neighbour nearby to the given field.